Separate registration of IDatasetForm plugins and blueprints
Blueprints need to be registered when there is an app available (ie
flask_app.py), but the actual IDatasetForm instances need to be readded
every time the environment is refreshed, as we reset them and start from
scratch. This happens eg when we explicitly load a plugin during tests
with `p.load()`.