Retrieving the package after plugins run
In the current implementation, the package is retrieved once, but not again after the plugin hooks. If a plugin modifies a package in the IResourceController.before_delete methods, the changes made to the package will be overwritten on line 191 when the package is updated using the cached package.

This change re-retrieves the package after the plugin method has run, so that any changes made by plugins are preserved.