[#4747] Pin Werkzeug
After upgrading Werkzeug to 0.15.2 we got the following failures:

        AssertionError: Environmental variable PATH_INFO is not a
        string: <type 'unicode'> (value: u'/ckan-admin/reset_config')

This is webtest complaining that the test request environ used in some
functional tests has a unicode value for environ['PATH_INFO']. Modern
versions of WebTest ensure that the url passed is a string, but sadly
the version we are stuck with only does so for GET requests (We are
stuck with this version of WebTest because newer versions have a higher
WebOb requirement which conflicts with our Pylons requirement).

We could ensure on our side that the urls provided to webtest are
strings but that would involve either

1. Change all app.post() calls in tests to app.post(str(url))
2. Change url_for itself to always return a sting, which obviously is
not a great solution.

Having Werkzeug pinned is not ideal though so we either find a
workaround or speed Pylons removal (allowing us to upgrade
webob/webtest)