[#4878] Allow custom CKAN callback URL for the DataPusher
The DataPusher pings back CKAN when performing or finishing
a job (calling the datapusher_hook action), and it does so
via an HTTP request to the host defined in ckan.site_url.

There are a number of scenarios where this does not work, eg:

* CKAN and DataPusher sitting behind a Firewall that doesn't
  allow external requests
* Standard Docker compose setup for development where the
  ckan.site_url is http://localhost:5000 or similar

This change adds a new config option that allows to define an
alternative internal URL that DataPusher can reach.