Add requirements.txt for python 3
Also:
* Install-by-source instructions is for python 3
* Adds initial python2->3 upgrade instructions
* Initial adjustments to ckan/common.py for import in python3

I'm keen to do 'trunk based development' for python 3 work now. So python 3 is the default in the docs and requirements.txt. I realize that the devs who use master for other work might be a bit surprised/disrupted by these changes. However I think it is worth it because the focus now needs to be python3, and it's better to switch master's default now rather than just before release.