Fix helper for getting user object in Python 3
The authz.py module has a function called `_get_user()` that returns the
user object from `g` if present or the DB one if exists. When outside
the context of a web request we had been defaulting to the Pylons `c`
object, which raises a `TypeError` if it's not ready. With Pylons gone,
we are asking to Flask's `g` if the user obj is there. `g` raises a
`RuntimeError` if it's not ready (if you are outside of an application
context in Flask terms), so I updated the helper.

Also removed duplication and updated the code to use `g`, which is what
we should be getting used to from now on.