Reinstate webtest dep for py2 to avoid WebOb conflict
We switched from nose to pytets, including for py2, so we don't use
WebTest any more. However we still use pylons in py2 and pylons requires WebTest>=1.1
so it installs WebTest2.0.34 but that in turn needs WebOb>=1.2, when we
specify WebOb==1.0.8, so there is a conflict and error message. So
that's why we need to continue to specify the WebTest version.
```
$ pip install -r requirements-py2.txt
...
Collecting WebTest>=1.1
  Downloading WebTest-2.0.34-py2.py3-none-any.whl (32 kB)
...
ERROR: webtest 2.0.34 has requirement WebOb>=1.2, but you'll have webob 1.0.8 which is incompatible.
```
I don't know why pip doesn't work this out! This error can be reproduced with
a blank virtualenv and:

    pip install pylons==0.9.7 WebOb==1.0.8