Add the g object to toolkit
Since CKAN 2.7 we have been favouring the `g` object in CKAN core in
favour of `c` to align with the Flask naming (`c` and `g` are just
aliases of the same object). But extensions can not
follow that pattern because `g` is not in the toolkit. This commit adds
it to the toolkit and marks `c` as deprecated.