plugin_extras support at the logic layer
Add a new plugin_extras field to the schema, ignored if you are not a
sysadmin and that expects a valid JSON object / dict.

For user_show, this field is not returned unless the erquester is a
sysdmin and the include_plugin_extras param is True. user_create and
user_update return the field if it was part of the input data (and you
are a sysadmin of course)