Add 'membership' explicitly to ROLE_PERMISSIONS
There are a couple of auth checks performed by
`has_user_permission_for_group_or_org` that use the "membership"
permission. This permission is not actually present in the
ROLE_PERMISSIONS mapping but the function defaults to checking "admin"
which has essentially the same effect. To be consistent this adds the
"membership" perm to the mapping.