[#5698] Filter activities from user at the database level
Fixes #5698

The `ckan.hide-activity-from-users` configuration option allows you to define
the users that you don't want to show activities from on the various activity
feeds. It defaults to the internal site user as this is generally used to
perform bulk internal processes like harvesting and you don't want to flood
the activity stream with their operations.

The problem is that we first get the relevant activities from the
database and then filter these activities in Python based on the users
in the config option.

This means that if more activities from the site user were performed
than the limit requested, they are fltered out and you get an empty
feed, eventhough the activities are still there.

These moves the filtering to the model queries.