Support for setting host and port on the ini file
It's useful to be able to set these on your local ini file and not have
to provide them each time when running `ckan run`. Paster use to support
this in the `[server:main]` section of the ini file.

Added `ckan.devserver.host` and `ckan.devserver.port` support, plus some
value checks for the port.