Revert #6954, check for a null session
In #6954 we added a custom check at the end of each request to decide
whether a session was null or not at then end of a request, to avoid
storing unneeded session files (Beaker always adds some keys to the
session regardless of whethere something is actually stored in it by
the app). With the CSRF protection implemented (#6863), the CSRF Flask
plugin always add a `_csrf_token` to the session so we do actually need
to store it.

The original issue described in #6954 was problematic because it meant
that session files were created on each different request, regardless of
the user. However, this does not happen in CKAN 2.10, probably because
of better user session management by Flask login, so we get the expected
result: session files are stored per user (or rather visitor or browser
session, as they are also stored on anonymous requests)