[#7028] Allow to shut down cookie based auth for API entirely
Refs #7028

Site maintainers can choose to completely ignore cookie based
authentication for API calls. When
`ckan.auth.disable_cookie_auth_in_api` is set to False, if there is a
user identified via the cookie and it is an API request,
it will be ignored entirely and CKAN will fall back to token based authentication.

This will probably break some JS widgets and view plugins from the
frontend so it should be used with caution.