[#7028] Consolidate all functional tests to environ overrides
If we mock the logged in user directly, the CSRF validation fails.
Using environ_overrides with {"REMOTE_USER": name} or {"Authorization":
token} bypasses the CSRF protection.

In these cases we need to use the REMOTE_USER variant because the
Authorization header is not passed to redirects, so we can't get the
final request content to test it