Replace usage of REMOTE_USER for authenticating requests
REMOTE_USER does not skip the CSRF protection like passing
`Authorization` does. It seems to do so when running multiple tests,
because previous ones have set the CSRF token in the session, but if you
run the test individually it fails. Note that if your test is testing a
redirect response you need to pass `HTTP_AUTHORIZATION` instead
otherwise the second request is not authenticated.