Allow ckan db upgrade to run against an old database
User.last_active was introduced in CKAN 2.10. When trying to run `ckan
db ugprade` against a database that was created before that you get the
following error:

    sqlalchemy.exc.ProgrammingError: (psycopg2.errors.UndefinedColumn) column user.last_active does not exist
    LINE 1: ...ey AS user_reset_key, "user".about AS user_about,
     "user".las..

This is caused by the `get_site_user` action call that we do on
`update_config()`. The changes detect this error and continue the
execution, similarly to how we check about a potential race condition.