Sanitize indexes in tuple keys when flattening dicts
When we "flatten" data dicts in the dictization layer we transform keys
like `extras__0__key` and `extras__0__value` to `('extras', 0, 'key')`
and `('extras', 0, 'key')`. As these keys are external inputs (generally
sent by the UI dataset form, but can also be sent directly with a
crafted POST request) we need to perform some validation. More
specifically:

* They are all ints
* They are all sequential, starting from 0

We keep the original order whenever possible, even if the input keys are
not sequential.

Added missing tests.