[8482] Fix permissions for groups_available and organizations_available
When checking access to the action organization_list_for_user, we should
be checking the current user's access, not the access of the user whose
groups or organizations we are enquiring about. The current user's name should be included in the context dict. The id of the user we are enquiring about should be included in the data_dict.