Fixes in the PyPI publish workflow
I separated the PyPI and Test PyPI worflows in two separate files (and
also two different GitHub environments) for ease of configuration and
security. Added better event checks (on tag push and push to master)
The publishing to Test PyPI on each merge requires another step to add a
timestamp to the version number otherwise the upload fails.