Remove CSRF check exemption for plugin blueprints
CSRF protection was introduced in 2.10.0 (February 2023). At the time we
allowed blueprints from plugins to skip the CSRF checks to allow time to
migrate their forms (basically adding the `{{ h.csrf_input() }}`
snippet). A config option `ckan.csrf_protection.ignore_extensions`
allowed to turn this exemption on and off, but now it feels like a good
time to enforce this across all extensions.